requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.5
azure-storage-blob==12.19.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
coloredlogs==15.0.1
comm==0.2.0
cryptography==41.0.5
dataclasses-json==0.6.2
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
docarray==0.39.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
filelock==3.13.1
flask==3.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jupyter-client==8.6.0
jupyter-core==5.5.0
kubernetes==28.1.0
langchain==0.0.335
langsmith==0.0.63
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.2
openai==1.2.3
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pdfminer.six==20221105
pexpect==4.8.0
platformdirs==4.0.0
posthog==3.0.2
prompt-toolkit==3.0.40
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
Pygments==2.16.1
pypdf==3.17.0
pypdf2==3.0.1
PyPika==0.48.9
pysqlite3-binary==0.5.2.post1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
safetensors==0.4.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
typer==0.9.0
types-requests==2.31.0.10
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.9
websocket-client==1.6.4
websockets==12.0
werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
